EstudianteBack/requirements.txt
Django==3.2.5
djangorestframework==3.12.2
mysql-connector-python==8.0.23
mysqlclient==2.0.3
PyMySQL==1.0.2
django-cors-headers==3.7.2